Move quorum intersection checker from scp to herder.
