+++ Builds/VisualStudio/stellar-core.vcxproj
@@ -303,8 +303,8 @@ exit /b 0
     <ClCompile Include="..\..\src\herder\TxSetFrame.cpp" />
     <ClCompile Include="..\..\src\herder\Upgrades.cpp" />
     <ClCompile Include="..\..\src\historywork\BatchDownloadWork.cpp" />
-    <ClCompile Include="..\..\src\scp\QuorumIntersectionCheckerImpl.cpp" />
-    <ClCompile Include="..\..\src\scp\test\QuorumIntersectionTests.cpp" />
+    <ClCompile Include="..\..\src\herder\QuorumIntersectionCheckerImpl.cpp" />
+    <ClCompile Include="..\..\src\herder\test\QuorumIntersectionTests.cpp" />
     <ClCompile Include="..\..\src\work\BatchWork.cpp" />
     <ClCompile Include="..\..\src\historywork\DownloadBucketsWork.cpp" />
     <ClCompile Include="..\..\src\historywork\FetchRecentQsetsWork.cpp" />
@@ -600,8 +600,8 @@ exit /b 0
     <ClInclude Include="..\..\src\herder\TxSetFrame.h" />
     <ClInclude Include="..\..\src\herder\Upgrades.h" />
     <ClInclude Include="..\..\src\historywork\BatchDownloadWork.h" />
-    <ClInclude Include="..\..\src\scp\QuorumIntersectionChecker.h" />
-    <ClInclude Include="..\..\src\scp\QuorumIntersectionCheckerImpl.h" />
+    <ClInclude Include="..\..\src\herder\QuorumIntersectionChecker.h" />
+    <ClInclude Include="..\..\src\herder\QuorumIntersectionCheckerImpl.h" />
     <ClInclude Include="..\..\src\work\BatchWork.h" />
     <ClInclude Include="..\..\src\historywork\DownloadBucketsWork.h" />
     <ClInclude Include="..\..\src\historywork\FetchRecentQsetsWork.h" />
