+++ src/herder/HerderImpl.cpp
@@ -9,6 +9,7 @@
 #include "herder/HerderPersistence.h"
 #include "herder/HerderUtils.h"
 #include "herder/LedgerCloseData.h"
+#include "herder/QuorumIntersectionChecker.h"
 #include "herder/TxSetFrame.h"
 #include "ledger/LedgerManager.h"
 #include "ledger/LedgerTxn.h"
@@ -21,7 +22,6 @@
 #include "main/PersistentState.h"
 #include "overlay/OverlayManager.h"
 #include "scp/LocalNode.h"
-#include "scp/QuorumIntersectionChecker.h"
 #include "scp/Slot.h"
 #include "transactions/TransactionUtils.h"
 #include "util/Logging.h"
