+++ src/herder/QuorumIntersectionCheckerImpl.cpp
