+++ src/herder/QuorumIntersectionCheckerImpl.h
